Remove UB in heapless::Vec::truncate()
The vector `truncate` method works by taking elements
of the vector past the truncation point and applying
`core::ptr::drop_in_place` to them.  It does this from
the end of the vector towards the front; carefully
decrementing the vector's length _before_ the drop in
order to ensure that a `panic!()`ing drop won't result
in a double drop.

However, to retrieve the element, `truncate` calls,
`self.as_mut_slice()`, which returns a slice of length
`len`: the _new_ length of the slice reflecting the
decrement.  Truncate then invokes UB by indexing the
element just past the end of the slice by calling,
`get_unchecked_mut(len))` on the slice.  We have
observed this generate a ud2 on x86_64.

The fix is to crib from the implementation of
`std::vec::Vec::truncate`, which avoids this problem
by creating the slice directly, as well as dropping
a subslice of elements, instead of in a loop.

Signed-off-by: Dan Cross <cross@oxidecomputer.com>